refactor: vehicle types
