+++ src/lib/api/vehicles.ts
@@ -2,6 +2,15 @@ import { backend } from '@/lib/services/backendService';
 
 import type { CreateVehicleParams, UpdateVehicleParams, Vehicle } from './vehicles.d';
 
+export type { Vehicle };
+
+export enum VehicleEnum {
+  TRUCK = 'TRUCK',
+  TRAILER = 'TRAILER',
+  SOLO_TRUCK = 'SOLO_TRUCK',
+  VAN = 'VAN',
+}
+
 export async function getVehicles() {
   return backend.get<Vehicle[]>('/api/fleet');
 }
