+++ src/lib/utils/vehicles.ts
@@ -1,0 +1,8 @@
+import { VehicleEnum } from '@/lib/api';
+
+export const vehicleTypeToPathMap = {
+  [VehicleEnum.TRUCK]: 'trucks',
+  [VehicleEnum.TRAILER]: 'trailers',
+  [VehicleEnum.SOLO_TRUCK]: 'solo-trucks',
+  [VehicleEnum.VAN]: 'vans',
+};
