+++ src/pages-components/Dashboard/FleetPage/Trucks/TrucksPage/TrucksTable.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link';
 import { useMemo } from 'react';
 
 import { VehicleAlertTooltip } from '@/components/alerts/VehicleAlertTooltip';
-import { type Vehicle, VehicleEnum } from '@/lib/api/vehicles.d';
+import { type Vehicle, VehicleEnum } from '@/lib/api/vehicles';
 import { getDataPointDateString } from '@/lib/utils/date';
 import { Box, FlexLayout, Table, Text } from '@/ui';
